Revert "Cbv 104 implement node api calls"
