+++ Back-End/src/dataSaver.py
@@ -1,4 +1,3 @@
-import datetime
 import pymongo
 from typing import Final
 
@@ -15,35 +14,6 @@ class dataSaver:
 
     def saveCANLocal(self, canBus):
         ...
-
-    def receiveTraffic(self, projectId, dbc, bus):
-
-        _msg = bus.recv()
-        _msgInfo = (dbc.get_message_by_frame_id(_msg.arbitration_id))
-        _msgData = str(dbc.decode_message(_msg.arbitration_id, _msg.data))
-
-        _myClient = pymongo.MongoClient(localDB)
-        _myDB = _myClient["TestPDB"]
-        _myCol = _myDB["TestColNodes"]
-
-        # Checks if node is in testCol_Nodes
-        if _myCol.find_one({'nodeID': str(_msg.arbitration_id)}) == None:
-            node =    {'projectId': projectId,
-                    'nodeID': str(_msg.arbitration_id),
-                    'name': str(_msgInfo.comment),
-                    'data': None,
-                    'position': None,
-                    'relationships': []}
-
-            self.storeNodes([node])
-
-        packet =    {'projectId': projectId,
-                    'timestamp': str(datetime.fromtimestamp(_msg.timestamp))[:-3],
-                    'type': str(_msg.dlc),
-                    'nodeId': str(_msgInfo.comment),
-                    'data': _msgData} 
-        self.storePackets([packet])
-        return
 
     def update(projectID, baudRate, initials, eventName, dbcFile, blacklistFile, archive):
         _myClient = pymongo.MongoClient(localDB)
@@ -110,20 +80,12 @@ class dataSaver:
 
         _myCol.delete_many({})
 
-    def storeNodes(nodes):
+    def storeNodes(node):
         _myClient = pymongo.MongoClient(localDB)
-        _myDB = _myClient["TestDB"]
+        _myDB = _myClient["TestPDB"]
         _myCol = _myDB["TestColNodes"]
 
-        _myCol.insert_many(nodes)
-
-    def updateNodes(projectID, updatedNodeList):
-        _myClient = pymongo.MongoClient(localDB)
-        _myDB = _myClient["TestDB"]
-        _myCol = _myDB["TestColNodes"]
-
-        for node in updatedNodeList:
-            _myCol.update_one({"projectId": projectID, "nodeId": node["nodeId"]}, {"$set": {"data": node["data"], "name": node["name"], "position": node["position"], "relationship": node["relationship"]}})
+        _myCol.insert_many(node)
 
 
 # This is meant for testing purposes only, in order to allow the quick and
