+++ Back-End/src/main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from routers import projectManager, packetManager, nodeManager
+from routers import projectManager, packetManager
 from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI()
@@ -16,7 +16,6 @@ app.add_middleware(
 
 app.include_router(projectManager.router)
 app.include_router(packetManager.router)
-app.include_router(nodeManager.router)
 
 @app.get("/")
 async def root():
